% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mipred.R, R/glm_mipred_old.R
\name{.expit}
\alias{.expit}
\title{Expit function converting odds to probability}
\usage{
.expit(x)

.expit(x)
}
\arguments{
\item{x}{Probability vector}

\item{x}{Probability vector}
}
\description{
Expit function converting odds to probability

Expit function converting odds to probability
}
