% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.mipred.R
\name{.glm.mipred.cmb1}
\alias{.glm.mipred.cmb1}
\title{Generalized linear model prediction using multiple imputation - prediction-averaging method}
\usage{
.glm.mipred.cmb1(formula, family, dataset, newdata, nimp, folds, miop)
}
\arguments{
\item{formula}{Formula used by fitting and prediction method}

\item{family}{Error distribution also determining the link function used}

\item{dataset}{A data frame containing calibration data}

\item{newdata}{A dataframe containing observations to be predicted}

\item{nimp}{Number of imputations for each observation}

\item{folds}{Number of folds defined in newdata}

\item{miop}{Mice options}
}
\description{
Generalized linear model prediction using multiple imputation - prediction-averaging method
}
