% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mipred.R
\name{.glm_mipred_cmb1}
\alias{.glm_mipred_cmb1}
\title{Binary prediction using multiple imputation - prediction-averaging method}
\usage{
.glm_mipred_cmb1(formula, family, dataset, newdata, nimp, folds, miop)
}
\arguments{
\item{formula}{Formula used by fitting and prediction method}

\item{family}{Error distribution also determining the link function used}

\item{dataset}{A data frame containing calibration data}

\item{newdata}{A dataframe containing observations to be predicted}

\item{nimp}{Number of imputations for each observation}

\item{folds}{Number of folds defined in newdata}

\item{miop}{Mice options}
}
\description{
Binary prediction using multiple imputation - prediction-averaging method
}
