% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mipred_cv.R
\name{.glm_mipred_cmb2_cv}
\alias{.glm_mipred_cmb2_cv}
\title{Cross-validation of binary prediction using multiple imputation - Rubin's rule coefficient-averaging method}
\usage{
.glm_mipred_cmb2_cv(formula, family, dataset, nimp, folds)
}
\arguments{
\item{formula}{Formula used by fitting and prediction method}

\item{family}{Error distribution also determining the link function used}

\item{dataset}{A data frame containing calibration data}

\item{nimp}{Number of imputations for each observation}

\item{folds}{The fold definition}
}
\description{
Cross-validation of binary prediction using multiple imputation - Rubin's rule coefficient-averaging method
}
