% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipred_cv.R
\name{mipred.cv}
\alias{mipred.cv}
\title{Cross-validation prediction using multiple imputation}
\usage{
mipred.cv(formula, family, data, nimp, folds, method = "averaging")
}
\arguments{
\item{formula}{A formula object providing a symbolic description of the
prediction model to be fitted.}

\item{family}{Specification of an appropriate error distribution and link
function.}

\item{data}{A data.frame containing calibration data on \code{n} samples.
Variables declared in \code{formula} must be found in \code{data}.}

\item{nimp}{Number of imputations used in the prediction of each observation.}

\item{folds}{Number of fold-partitions used in cross-validation. Put \code{folds=n}
for leave-one-out cross-validation.}

\item{method}{Imputation combination method. This defaults to
\code{"averaging"} for the prediction-averaging approach. The alternative
\code{"rubin"} applies the Rubin's rules pooled model.}
}
\value{
A list containing predictions.
\describe{\item{\code{pred}}{Matrix of predictions on the scale of the response
variable of dimension \code{n} by \code{nimp}.}
\item{\code{linpred}}{Matrix of predictions on the scale
  of the linear predictor of dimension \code{n} by \code{nimp}.}}
}
\description{
Calculates cross-validated predictions based on within-sample  assessment and calibration using
generalized linear models with multiple imputations to account for missing values
in predictor data.
}
\examples{
cll_bin<-cll # Generate a copy and construct binary outcome from survival information
cll_bin$srv5y_s[cll_bin$srv5y>12] <- 0  # Apply administrative censorship at t=12 months
cll_bin$srv5y[cll_bin$srv5y>12]  <- 12
cll_bin$Status[cll_bin$srv5y_s==1]<- 1  # Define the new binary "Status" outcome variable
cll_bin$Status[cll_bin$srv5y_s==0] <- 0  # As numeric -> 1:Dead, 0:Alive
cll_bin$Censor <- NULL # Remove survival outcomes
cll_bin$srv5y <- NULL
cll_bin$srv5y_s <- NULL

# Cross-validate prediction using logistic regression in the first 100 samples
# Apply prediction-averaging using 5 imputations and 5 folds
output<-mipred.cv(Status~perfstat+remstat+cyto, family=binomial, data=cll_bin[1:100,],
nimp=5, folds=5)

}
\references{
\url{https://arxiv.org/abs/1810.05099}
}
\seealso{
\code{\link{mice}}
}
\author{
Bart J A Mertens, \email{b.mertens@lumc.nl}
}
