% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.mipred.R
\name{.expit}
\alias{.expit}
\title{Expit function converting odds to probability}
\usage{
.expit(x)
}
\arguments{
\item{x}{Probability vector}
}
\value{
The expit transform of x (inverse logit)
}
\description{
Expit function converting odds to probability
}
\note{
This is an internal 'mipred' function and not intended to be called directly
}
