% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.mipred.cv.R
\name{.glm.mipred.cmb2.cv}
\alias{.glm.mipred.cmb2.cv}
\title{Cross-validation of generalized linear model prediction using multiple imputation - Rubin's rule coefficient-averaging method}
\usage{
.glm.mipred.cmb2.cv(formula, family, dataset, nimp, folds, miop)
}
\arguments{
\item{formula}{Formula used by fitting and prediction method}

\item{family}{Error distribution also determining the link function used}

\item{dataset}{A data frame containing calibration data}

\item{nimp}{Number of imputations for each observation}

\item{folds}{Number of folds defined in data}

\item{miop}{Mice options}
}
\description{
Cross-validation of generalized linear model prediction using multiple imputation - Rubin's rule coefficient-averaging method
}
