% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipred.R
\name{mipred}
\alias{mipred}
\title{Prediction using multiple imputation}
\usage{
mipred(formula, family, data, newdata, nimp, method = "averaging")
}
\arguments{
\item{formula}{A formula object providing a symbolic description of the prediction model to be fitted.}

\item{family}{Specification of an appropriate error distribution and link function.}

\item{data}{A data.frame containing calibration data on \code{n} samples. Variables declared in \code{formula} must be found in \code{data}.}

\item{newdata}{A data.frame containing the predictors for observations to be predicted on \code{m} samples. This must have the same structure and variables as \code{data}, except for the outcome variable which is ignored in the construction of the predictions and can therefor be excluded from the object.}

\item{nimp}{Number of imputations used in the prediction of each observation.}

\item{method}{Imputation combination method. This defaults to \code{"averaging"} for the prediction-averaging approach. The alternative \code{"rubin"} applies the Rubin's rules pooled model.}
}
\value{
A list containing predictions.
\describe{
  \item{\code{pred}}{Matrix of predictions on the scale of the response variable of dimension \code{m} by \code{nimp}.}
  \item{\code{linpred}}{Matrix of predictions on the scale of the linear predictor of dimension \code{m} by \code{nimp}.}
}
}
\description{
Calculates predictions from generalized linear models when multiple imputations are used to
account for missing values in predictor data.
}
\examples{

load("cll.rda")
cll_bin<-cll # Generate a copy and construct binary outcome from survival information
cll_bin$srv5y_s[cll_bin$srv5y>12] <- 0  # Apply administrative censorship at t=12 months
cll_bin$srv5y[cll_bin$srv5y>12]  <- 12
cll_bin$Status[cll_bin$srv5y_s==1]<- 1  # Define the new binary "Status" outcome variable
cll_bin$Status[cll_bin$srv5y_s==0] <- 0  # As numeric -> 1:Dead, 0:Alive
cll_bin$Censor <- NULL # Remove survival outcomes
cll_bin$srv5y <- NULL
cll_bin$srv5y_s <- NULL

set.seed(12345)
# Predict observations 501 to 507 using the first 500 records to calibrate predictors
# Apply prediction-averaging using 5 imputations
output<-mipred(Status~age10+perfstat+remstat+cyto+asct+donor+sex_match+cond,family=binomial,data=cll_bin[1:500,], newdata=cll_bin[501:507,-10], nimp=5)

}
\references{
\url{https://arxiv.org/abs/1810.05099}
}
\seealso{
\code{\link{mice}}
}
\author{
Bart J A Mertens, \email{b.mertens@lumc.nl}
}
